Update to upstream version '25.1.0'
with Debian dir fea345e3e055f271700956e2337d1d7c93b8b8d2

+++ xkb/xkmread.c
@@ -49,7 +49,7 @@ XkbInternAtom(char *str, Bool only_if_exists)
 {
     if (str == NULL)
         return None;
-    return MakeAtom(str, strlen(str), !only_if_exists);
+    return MakeAtom(str, (unsigned int)strlen(str), !only_if_exists);
 }
 
 /***====================================================================***/
